+++ src/nfan/fan/sys/MapType.cs
@@ -31,7 +31,7 @@ namespace Fan.Sys
   // Type
   //////////////////////////////////////////////////////////////////////////
 
-    public override Int hash()
+    public override Long hash()
     {
       return signature().hash();
     }
