+++ src/testNative/net/NativeSubPeer.cs
@@ -36,12 +36,12 @@ namespace Fan.TestNative
       return Str.make("subNative working");
     }
 
-    public Int subfX(NativeSub self)
+    public Long subfX(NativeSub self)
     {
       return fX(self);
     }
 
-    public Int subGetPeerZ(NativeSub self)
+    public Long subGetPeerZ(NativeSub self)
     {
       return getPeerZ(self);
     }
